+++ spec/rubocop/cop/capybara/rspec/specific_matcher_spec.rb
@@ -163,6 +163,23 @@ RSpec.describe RuboCop::Cop::Capybara::RSpec::SpecificMatcher do
     end
   end
 
+  it 'registers an offense for abstract matcher when first argument ' \
+     'is input with replaceable field attributes' do
+    expect_offense(<<~RUBY)
+      expect(page).to have_no_css('input[placeholder="foo"]')
+                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer `have_no_field` over `have_no_css`.
+      expect(page).to have_no_css('input[type="checkbox"]')
+                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer `have_no_field` over `have_no_css`.
+    RUBY
+  end
+
+  it 'does not register an offense for input attribute modifiers' do
+    expect_no_offenses(<<~RUBY)
+      expect(page).to have_css('input[placeholder="foo" i]')
+      expect(page).to have_css('input[placeholder="foo" s]')
+    RUBY
+  end
+
   it 'registers an offense when using abstract matcher with ' \
      'first argument is element with multiple replaceable attributes' do
     expect_offense(<<~RUBY)