+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix CSS attribute parsing when attribute values contain quotes. ([@ydah])
 - Add support for using `find` with `:link` and `:field` in `Capybara/SpecificFinders` cop. ([@ydah])
 - Fix CSS attribute parsing to return raw values and format Ruby literals in autocorrection. ([@ydah])
+- Fix a false negative for `Capybara/SpecificFinders` when using `find` with input field attributes. ([@ydah])
 
 ## 3.0.0 (2026-06-22)